Move Caddyfile parsing into caddyfile.go
